feat: add metrics commit count and commit time.

+++ source/dnode/vnode/src/inc/vnodeInt.h
@@ -485,8 +485,8 @@ struct SVnodeWriteMetrics {
   int64_t memory_table_size;
   int64_t commit_count;
   int64_t merge_count;
-  double  commit_time;
-  double  merge_time;
+  int64_t commit_time;
+  int64_t merge_time;
   int64_t block_commit_time;
   int64_t memtable_wait_time;
 };
